[refactor] 패키지 변경, 안쓰는 클래스 제거
